tests/functional/requirements.txt
multidict==6.0.5
pytest==7.4.3
aiohttp==3.9.3
elasticsearch==8.6.0
pydantic==1.10.2
pytest-asyncio==0.23.6
python-dotenv==1.0.1
redis==5.0.4
orjson==3.10.3